requirements.txt
torch
torchvision
timm==0.3.2
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
# deepspeed==0.4.0
scipy
